--- result_1.json
@@ -1,1 +1,0 @@
-{"meta_data": {"sheet_size": {"width": 282.96, "height": 282.96}}, "body": {"GRIFFON AVIATION SERVICES LLC": {"value": "", "type_value": {"type": "string", "pattern": ""}, "type_key": {"type": "string", "pattern": "^[A-Z ]+$"}, "key_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "10.509"}, "val_font_info": {"font_name": "", "font_size": ""}, "separator": "", "coordinates": {"x0": "6.005", "y0": "264.969"}}, "PN": {"value": " tst", "type_value": {"type": "string", "pattern": "^[a-z]+$"}, "type_key": {"type": "string", "pattern": "^[A-Z]+$"}, "key_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "val_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "separator": ":", "coordinates": {"x0": "13.512", "y0": "233.327"}}, "DESCRIPTION": {"value": " PART", "type_value": {"type": "string", "pattern": "^[A-Z]+$"}, "type_key": {"type": "string", "pattern": "^[A-Z]+$"}, "key_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "val_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "separator": ":", "coordinates": {"x0": "13.512", "y0": "222.067"}}, "LOCATION": {"value": " 111", "type_value": {"type": "number", "pattern": "^\\d+$"}, "type_key": {"type": "string", "pattern": "^[A-Z]+$"}, "key_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "val_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "separator": ":", "coordinates": {"x0": "13.512", "y0": "210.807"}}, "RECEIVER#": {"value": " 9", "type_value": {"type": "number", "pattern": "^\\d+$"}, "type_key": {"type": "string", "pattern": "^[A-Z#]+$"}, "key_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "val_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "separator": ":", "coordinates": {"x0": "13.512", "y0": "199.547"}}, "EXP DATE": {"value": " 13.04.2022", "type_value": {"type": "date", "pattern": "%d.%m.%Y"}, "type_key": {"type": "string", "pattern": "^[A-Z ]+$"}, "key_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "val_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "separator": ":", "coordinates": {"x0": "13.512", "y0": "188.287"}}, "CERT SOURCE": {"value": " wef", "type_value": {"type": "string", "pattern": "^[a-z]+$"}, "type_key": {"type": "string", "pattern": "^[A-Z ]+$"}, "key_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "val_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "separator": ":", "coordinates": {"x0": "13.512", "y0": "177.027"}}, "SN": {"value": " 123123", "type_value": {"type": "number", "pattern": "^\\d+$"}, "type_key": {"type": "string", "pattern": "^[A-Z]+$"}, "key_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "val_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "separator": ":", "coordinates": {"x0": "141.5", "y0": "233.327"}}, "CONDITION": {"value": " FN", "type_value": {"type": "string", "pattern": "^[A-Z]+$"}, "type_key": {"type": "string", "pattern": "^[A-Z]+$"}, "key_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "val_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "separator": ":", "coordinates": {"x0": "141.5", "y0": "210.807"}}, "UOM": {"value": " EA", "type_value": {"type": "string", "pattern": "^[A-Z]+$"}, "type_key": {"type": "string", "pattern": "^[A-Z]+$"}, "key_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "val_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "separator": ":", "coordinates": {"x0": "141.5", "y0": "199.547"}}, "PO": {"value": " P101", "type_value": {"type": "string", "pattern": "^[A-Z0-9]+$"}, "type_key": {"type": "string", "pattern": "^[A-Z]+$"}, "key_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "val_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "separator": ":", "coordinates": {"x0": "141.5", "y0": "188.287"}}, "REC.DATE": {"value": " 18.04.2022", "type_value": {"type": "date", "pattern": "%d.%m.%Y"}, "type_key": {"type": "string", "pattern": ""}, "key_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "val_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "separator": ":", "coordinates": {"x0": "13.512", "y0": "165.767"}}, "MFG": {"value": " efwfe", "type_value": {"type": "string", "pattern": "^[a-z]+$"}, "type_key": {"type": "string", "pattern": "^[A-Z]+$"}, "key_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "val_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "separator": ":", "coordinates": {"x0": "141.5", "y0": "165.767"}}, "BATCH# ": {"value": " 1", "type_value": {"type": "number", "pattern": "^\\d+$"}, "type_key": {"type": "string", "pattern": "^[A-Z#]+$"}, "key_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "val_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "separator": ":", "coordinates": {"x0": "13.512", "y0": "154.507"}}, "REMARK": {"value": "", "type_value": {"type": "string", "pattern": ""}, "type_key": {"type": "string", "pattern": "^[A-Z]+$"}, "key_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "val_font_info": {"font_name": "", "font_size": ""}, "separator": ":", "coordinates": {"x0": "13.512", "y0": "143.247"}}, "TAGGED BY": {"value": " ", "type_value": {"type": "string", "pattern": ""}, "type_key": {"type": "string", "pattern": "^[A-Z ]+$"}, "key_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "val_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "separator": ":", "coordinates": {"x0": "13.512", "y0": "131.987"}}, "Qty": {"value": " 1", "type_value": {"type": "number", "pattern": "^\\d+$"}, "type_key": {"type": "string", "pattern": "^[a-zA-Z0-9]+$"}, "key_font_info": {"font_name": "BAAAAA+ArialMT", "font_size": "7.507"}, "val_font_info": {"font_name": "BAAAAA+ArialMT", "font_size": "7.507"}, "separator": ":", "coordinates": {"x0": "17.265", "y0": "87.698"}}, "DOM": {"value": " 13.04.2022", "type_value": {"type": "date", "pattern": "%d.%m.%Y"}, "type_key": {"type": "string", "pattern": "^[A-Z]+$"}, "key_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "val_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "separator": ":", "coordinates": {"x0": "141.5", "y0": "154.507"}}, "LOT# ": {"value": " 1", "type_value": {"type": "number", "pattern": "^\\d+$"}, "type_key": {"type": "string", "pattern": "^[A-Z#]+$"}, "key_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "val_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "separator": ":", "coordinates": {"x0": "141.5", "y0": "143.247"}}, "NOTES": {"value": "", "type_value": {"type": "string", "pattern": ""}, "type_key": {"type": "string", "pattern": "^[A-Z]+$"}, "key_font_info": {"font_name": "AAAAAA+Arial-BoldMT", "font_size": "7.507"}, "val_font_info": {"font_name": "", "font_size": ""}, "separator": ":", "coordinates": {"x0": "141.5", "y0": "131.987"}}, "inspection notes": {"value": "", "type_value": {"type": "string", "pattern": ""}, "type_key": {"type": "string", "pattern": ""}, "key_font_info": {"font_name": "BAAAAA+ArialMT", "font_size": "7.507"}, "val_font_info": {"font_name": "", "font_size": ""}, "separator": "", "coordinates": {"x0": "146.004", "y0": "119.226"}}}}
